// core/src/lib/camel/topology-types.ts
export type Node = RFLow.Node & {
  readonly data: {
    stepType: StepType;
    absolutePath: string;
    iconName: string;
    label: string;
    operation: StepType;
  };
};
export type Edge = RFLow.Edge;
// ...
export interface Step {
  steps?: Step[];
  choice?: {
    when?: Step[];
    otherwise?: Step;
  };
  split?: {
    steps?: Step[];
  };
  join?: {
    aggregationStrategyRef?: string;
  };
  aggregate?: {
    correlationExpression?: string;
    completionSize?: number;
    completionTimeout?: number;
    aggregationStrategyRef?: string;
    steps?: Step[];
  };
  circuitBreaker?: {
    resilience4jConfiguration?: Record<string, unknown>;
    steps?: Step[];
    onFallback?: {
      steps?: Step[];
    };
  };
  recipientList?: {
    expression?: Record<string, unknown>;
    delimiter?: string;
    parallelProcessing?: boolean;
    stopOnException?: boolean;
    steps?: Step[];
  };
  filter?: {
    expression?: string;
    steps?: Step[];
  };
  loop?: {
    expression?: Record<string, unknown>;
    copy?: boolean;
    doWhile?: boolean;
    steps?: Step[];
  };
  saga?: {
    propagation?: string;
    completionMode?: string;
    compensation?: string;
    completion?: string;
    steps?: Step[];
  };
  transform?: {
    expression?: string;
  };
  custom?: {
    componentType: string;
    properties?: Record<string, unknown>;
  };
  doTry?: {
    steps?: Step[];
    doCatch?: {
      steps?: Step[];
    };
    doFinally?: {
      steps?: Step[];
    };
  };
  multicast?: {
    parallelProcessing?: boolean;
    strategyRef?: string;
    stopOnException?: boolean;
    steps?: Step[];
  };
  loadBalance?: {
    strategyRef?: string;
    steps?: Step[];
  };
  log?: {
    message?: string;
    loggingLevel?: string;
  };
  camel?: {
    uri: string;
  };
  kafka?: {
    topic: string;
    brokers: string;
    clientId?: string;
    groupId?: string;
    autoOffsetReset?: string;
    additionalProperties?: Record<string, unknown>;
  };
  http?: {
    uri: string;
    method?: string;
    authenticationPreemptive?: boolean;
    authenticationUsername?: string;
    authenticationPassword?: string;
    headers?: Record<string, unknown>;
    queryParameters?: Record<string, unknown>;
    body?: string;
  };
  rest?: {
    verb: string;
  };
  soap?: {
    wsdlUri: string;
    serviceName?: string;
    portName?: string;
    operationName?: string;
    headers?: Record<string, unknown>;
    body?: string;
  };
  wsdl?: {
    uri: string;
    serviceName?: string;
    portName?: string;
    operationName?: string;
    headers?: Record<string, unknown>;
    body?: string;
  };
  add_step?: {};
  add_between?: {};
  when?: {};
  otherwise?: {};
  doCatch?: {};
  doFinally?: {};
  [key: string]: any;
}

export type Route = {
  route?: {
    id: string;
    nodePrefixId: string;
    from: {
      id: string;
      uri: string;
      steps?: Step[];
    };
  };
};

export type CamelConfig = {
  data: Route[];
  comments?: Record<string, string>;
};

export enum STEP_TYPE {
  CHOICE = "choice",
  SPLIT = "split",
  JOIN = "join",
  AGGREGATE = "aggregate",
  CIRCUIT_BREAKER = "circuitBreaker",
  RECIPIENT_LIST = "recipientList",
  FILTER = "filter",
  LOOP = "loop",
  SAGA = "saga",
  TRANSFORM = "transform",
  CUSTOM = "custom",
  CAMEL = "camel",
  KAFKA = "kafka",
  HTTP = "http",
  REST = "rest",
  SOAP = "soap",
  WSDL = "wsdl",
  ADD_STEP = "add-step",
  WHEN = "when",
  OTHERWISE = "otherwise",
  DO_CATCH = "doCatch",
  DO_TRY = "doTry",
  DO_FINALLY = "doFinally",
  CUSTOM_FALLBACK = "fallback",
  MULTICAST = "multicast",
  LOAD_BALANCE = "loadBalance",
  ADD_BETWEEN = "add-between",
  ADD_WHEN = "add-when",
  ADD_DO_CATCH = "add-doCatch",
  EDIT = "edit-step",
}

export type StepType = `${STEP_TYPE}`;